Extract lib-tags
