+++ libraries/CMakeLists.txt
@@ -31,6 +31,7 @@ set( LIBRARIES
    lib-project-history
    lib-graphics
    lib-audio-graph
+   lib-tags
 )
 
 if ( ${_OPT}has_networking )
