+++ src/export/ExportCL.cpp
@@ -34,7 +34,7 @@
 #include "Prefs.h"
 #include "../SelectFile.h"
 #include "../ShuttleGui.h"
-#include "../Tags.h"
+#include "Tags.h"
 #include "Track.h"
 #include "float_cast.h"
 #include "../widgets/FileHistory.h"
