+++ src/export/ExportFFmpegDialogs.cpp
@@ -63,7 +63,7 @@
 #include "../widgets/FileDialog/FileDialog.h"
 
 #include "Mix.h"
-#include "../Tags.h"
+#include "Tags.h"
 #include "../widgets/AudacityMessageBox.h"
 #include "../widgets/HelpSystem.h"
