+++ src/export/ExportFLAC.cpp
@@ -31,7 +31,7 @@ Joshua Haberman
 #include "Prefs.h"
 #include "../ShuttleGui.h"
 
-#include "../Tags.h"
+#include "Tags.h"
 #include "Track.h"
 
 #include "../widgets/AudacityMessageBox.h"
