+++ src/export/ExportMP2.cpp
@@ -52,7 +52,7 @@
 #include "Prefs.h"
 #include "ProjectRate.h"
 #include "../ShuttleGui.h"
-#include "../Tags.h"
+#include "Tags.h"
 #include "Track.h"
 #include "../widgets/AudacityMessageBox.h"
 #include "../widgets/ProgressDialog.h"
