+++ src/export/ExportPCM.cpp
@@ -30,7 +30,7 @@
 #include "Prefs.h"
 #include "ProjectRate.h"
 #include "../ShuttleGui.h"
-#include "../Tags.h"
+#include "Tags.h"
 #include "Track.h"
 #include "../widgets/AudacityMessageBox.h"
 #include "../widgets/ProgressDialog.h"
