+++ src/import/ImportMP3_MPG123.cpp
@@ -27,7 +27,7 @@
 #include <mpg123.h>
 
 #include "Prefs.h"
-#include "../Tags.h"
+#include "Tags.h"
 #include "../WaveTrack.h"
 #include "../widgets/AudacityMessageBox.h"
 #include "../widgets/ProgressDialog.h"
